# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reeve",
    "description": "Fleet desired-state manager (Margo-inspired). Human API consumed by the reeve UI, plus the device-facing wire surface.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/devices/{device_id}/journal": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "GET /api/devices/{device_id}/journal (viewer+) — status journal\npage, newest first (§7.3 forensic record; late-backfilled records\nappear in seq order, not arrival order).",
        "operationId": "journal",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "Device id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size; default 100, max 1000",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "before_seq",
            "in": "query",
            "description": "Return records with seq strictly below this (older page)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One journal page, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below viewer role"
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reeve/v1/capabilities": {
      "get": {
        "tags": [
          "device"
        ],
        "summary": "GET /api/reeve/v1/capabilities (device-auth'd; anonymous pull of\nanything device-facing is disabled by default, §10.2).",
        "operationId": "capabilities",
        "responses": {
          "200": {
            "description": "Server extension advertisement (spec/reeve/01-framework.md §3.3)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerCapabilities"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
    "/api/reeve/v1/manifest": {
      "get": {
        "tags": [
          "device"
        ],
        "summary": "GET /api/reeve/v1/manifest — the device-scoped State Manifest\n(§10.2). Renders on demand when the device's row is behind the\nlocal head (fresh enrollment, missed pass), then serves the stored\nmanifest bytes verbatim so the ETag is exact.",
        "operationId": "manifest",
        "parameters": [
          {
            "name": "if-none-match",
            "in": "header",
            "description": "RFC 9110 conditional GET against the manifest's strong ETag",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The device's current State Manifest (ETag header set)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateManifest"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match matched)"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Enrolled but never rendered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppManifestEntry": {
        "type": "object",
        "description": "Per-app State Manifest entry. The agent diffs `secrets_version`\nseparately from the bundle digest: bundle unchanged +\n`secrets_version` changed ⇒ re-resolve secrets and minimally\nre-up, no bundle re-pull (spec/reeve/10-secrets.md §12;\ndocs/decisions/agent.md).",
        "required": [
          "appId"
        ],
        "properties": {
          "appId": {
            "type": "string",
            "description": "Application id (matches the rendered app dir /\n`ApplicationDeployment.spec.applicationId`)."
          },
          "deploymentId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Deterministic deployment id of the rendered\n`ApplicationDeployment` (docs/decisions/tree-render.md)."
          },
          "secrets_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hash of resolved secret names+versions, never values\n(spec/reeve/10-secrets.md §12; docs/decisions/delivery.md D13).\nSpelled `secrets_version` on the wire — the exact token used\nnormatively throughout spec/reeve/."
          }
        }
      },
...
      "BundleRef": {
        "type": "object",
        "description": "Render-bundle reference — shaped after Margo's\n`DeploymentBundleRef` (`workload-management-api-1.0.0.yaml`):\nmediaType, digest, advisory sizeBytes, content-addressable url.",
        "required": [
          "digest",
          "url"
        ],
        "properties": {
          "digest": {
            "type": "string",
            "description": "`sha256:<hex>`; MUST be verified after pull — `sizeBytes` is\nadvisory only, never integrity."
          },
          "mediaType": {
            "type": [
              "string",
              "null"
            ]
          },
          "sizeBytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "url": {
            "type": "string",
            "description": "Content-addressed pull URL (native OCI /v2 routes,\ndocs/decisions/delivery.md D7)."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The one error-body shape every JSON error response uses:\n`{ \"error\": \"<human-readable message>\" }`. Schema component for the\nD10 openapi.json (docs/decisions/ui.md); handlers keep building the\nbody via `json!` — this type documents the contract.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "JournalEntry": {
        "type": "object",
        "description": "One status-journal record (spec/reeve/05-health-journal.md §7.3).",
        "required": [
          "seq",
          "observedAt",
          "receivedAt",
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "`status` | `health` | `lifecycle` | `gap`."
          },
          "observedAt": {
            "type": "string",
            "description": "Device-assigned original RFC 3339 timestamp, verbatim."
          },
          "payload": {
            "description": "The journaled payload; absent for gap marks."
          },
          "receivedAt": {
            "type": "integer",
            "format": "int64",
            "description": "Server receipt time, unix seconds."
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "JournalPage": {
        "type": "object",
        "description": "`GET /api/devices/{device_id}/journal` response: newest first.",
        "required": [
          "records"
        ],
        "properties": {
          "nextBeforeSeq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Pass as `before_seq` to fetch the next (older) page; absent\nwhen this page reached the start of the journal."
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntry"
            }
          }
        }
      },
...
      "ManifestVersion": {
        "type": "integer",
        "format": "int64",
        "description": "`manifestVersion` — logically the pair `(epoch, counter)` compared\nlexicographically, encoded on the wire as ONE monotonically\nincreasing unsigned 64-bit integer: epoch in the high 16 bits,\ncounter in the low 48 (spec/reeve/08-packaging.md §10.2). Plain\ninteger comparison IS the pair comparison, and the wire value\nstays exactly Margo's modeled shape (`ManifestVersion`: monotonic\nu64 in `[1, 2^64-1]`, `workload-management-api-1.0.0.yaml` —\n\"The first manifest MUST use 1\").\n\nAgent rules (§10.2): a non-increasing value is rejected and logged\nas a SECURITY event; an increase that bumps the epoch bits is\naccepted and logged as a NOTABLE event (a restore happened,\nspec/reeve/07-durability.md §9.5 restore fencing).",
        "minimum": 0
      },
...
      "ServerCapabilities": {
        "type": "object",
        "description": "The server-side advertisement at\n`GET /api/reeve/v1/capabilities` — \"its extension list plus\nserver version\" (spec/reeve/01-framework.md §3.3). 404 or any\nerror means \"vanilla Margo server\"; the agent proceeds with pure\nMargo behavior.",
        "required": [
          "serverVersion"
        ],
        "properties": {
          "extensions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Same `\"rev-NNN/V\"` grammar as the agent side."
          },
          "serverVersion": {
            "type": "string"
          }
        }
      },
...
      "StateManifest": {
        "type": "object",
        "description": "The reeve State Manifest body returned by\n`GET /api/reeve/v1/manifest` (spec/reeve/08-packaging.md §10.2) —\nState-Manifest-shaped after Margo's `UnsignedAppStateManifest`\n(`workload-management-api-1.0.0.yaml`): `manifestVersion`,\nrender-bundle digest + pull URL, per-app `secrets_version`.",
        "required": [
          "manifestVersion"
        ],
        "properties": {
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppManifestEntry"
            },
            "description": "Per-app entries carrying `secrets_version`\n(docs/decisions/delivery.md D13; spec/reeve/10-secrets.md §12)."
          },
          "bundle": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleRef",
                "description": "Render-bundle reference. Follows Margo's `DeploymentBundleRef`\nnull rule: with zero apps the property MUST be present with\nthe value `null`, never omitted — so it is always serialized."
              }
            ]
          },
          "manifestVersion": {
            "$ref": "#/components/schemas/ManifestVersion"
          }
        }
      },
...
    }
  },
...
}